refactor(#451): organize question views
